deleted warning..
git-svn-id: https://svn.code.sf.net/p/wired/code/branches/saveXml@1171 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/gui/Pattern.cpp
@@ -213,7 +213,8 @@ void					Pattern::OnMotion(wxMouseEvent &e)
 	{
 	  if (SeqPanel->SelectedItems.size() == 1)
 	    {
-	      trackto = TrackIndex + floor((e.GetPosition().y  * SeqPanel->VertZoomFactor) / TRACK_HEIGHT);
+	      trackto = (int)(TrackIndex + floor((e.GetPosition().y  * SeqPanel->VertZoomFactor)
+						   / TRACK_HEIGHT));
 	      if (SeqPanel->IsAudioTrack(TrackIndex) == SeqPanel->IsAudioTrack(trackto))
 		{
 		  SeqPanel->ChangeMouseCursor(wxCursor(wxCURSOR_HAND));
